requirements.txt
asgiref==3.5.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.11.0
Django==4.0.4
django-calculation==0.0.10
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-pagination==1.0.7
django-qr-code==3.0.0
djangorestframework==3.13.1
fonttools==4.32.0
gunicorn==20.1.0
idna==2.10
kiwisolver==1.4.2
matplotlib==3.5.1
nodeenv==1.6.0
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-orange-sms==0.1.4
pytz==2022.1
requests==2.25.1
segno==1.4.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.4
whitenoise==6.0.0
